+++ product_code/user_service/api/src/main/scala/de/upb/cs/uc4/user/api/UserService.scala
@@ -9,9 +9,8 @@ import com.lightbend.lagom.scaladsl.api.{ Descriptor, Service, ServiceAcl, Servi
 import de.upb.cs.uc4.shared.client.UC4Service
 import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 import de.upb.cs.uc4.shared.client.message_serialization.CustomMessageSerializer
-import de.upb.cs.uc4.user.model.post.PostMessageUser
 import de.upb.cs.uc4.user.model.user.{ Admin, Lecturer, Student, User }
-import de.upb.cs.uc4.user.model.{ GetAllUsersResponse, JsonRole, MatriculationUpdate }
+import de.upb.cs.uc4.user.model._
 
 /** The UserService interface.
   *