+++ product_code/user_service/CHANGELOG.md
@@ -1,7 +1,15 @@
+# [WIP - v0.13.2](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.13.1...user-v0.13.2) (2020-XX-XX)
+## Feature
+## Refactor
+ - Changed PostMessageUser to use proper inheritance
+## Bugfix
+
+
 # [v0.13.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.13.0...user-v0.13.1) (2020-11-25)
 ## Feature
  - Added governmentId to PostMessageUser
  - Added enrollmentIdSecret to user object
+ - Changed default users accordingly, default Users govId is "governmentIdStudent", enrollmentIdSecret is base64 encoding of "studentstudent", "lecturerlecturer" or "adminadmin"
 ## Refactor
 ## Bugfix
 
@@ -19,7 +27,6 @@
 
 # [v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.11.0...user-v0.11.1) (2020-11-06)
 ## Feature
- - Changed default users accordingly, default Users govId is "governmentIdStudent", enrollmentIdSecret is base64 encoding of "studentstudent", "lecturerlecturer" or "adminadmin"
  - Added support for ETags
 ## Refactor
  - Changed email addresses, first names and last names for the default users